+++ examples/gjs/gio-2-cat-promisify/main.ts
@@ -9,7 +9,7 @@
  */
 
 import Gio from "gi://Gio?version=2.0";
-const ByteArray = imports.byteArray;
+const textDecoder = new TextDecoder();
 
 Gio._promisify(Gio.File.prototype, 'load_contents_async', 'load_contents_finish');
 
@@ -17,7 +17,7 @@ async function cat(filename: string) {
     const file = Gio.file_new_for_path(filename);
 
     const [contents] = await file.load_contents_async(null);
-    print(ByteArray.toString(contents));
+    print(textDecoder.decode(contents));
 }
 
 if (ARGV.length !== 1)