+++ packages/generator-typescript/templates/gjs/module-ambient.d.ts
@@ -1,7 +1,7 @@
 <%#
-    This EJS template is used for the generated .d.ts files for ambient typescript module support.
-    See https://www.typescriptlang.org/docs/handbook/modules.html#ambient-modules
-%>
+    // This EJS template is used for the generated .d.ts files for ambient typescript module support.
+    // See https://www.typescriptlang.org/docs/handbook/modules.html#ambient-modules
+_%>
 <%_ const pkg = dep.get(girModule.namespace, girModule.version) _%>
 <%_ let moduleImportStr = ""; _%>
 <%_ if(package){ _%>
@@ -23,7 +23,8 @@ declare module 'gi://<%= name %>?version=<%= version %>' {
     export default <%- girModule.importNamespace -%>;
 }
 
-<%_ if(!dep.hasConflict(name, version)){ _%>
+<%# // Generate ambient module declarations Without version number if there are no conflicts or the target is an NPM package _%>
+<%_ if(package || !dep.hasConflict(name, version)){ _%>
 declare module 'gi://<%= name %>' {
     <%- moduleImportStr %>;
     export default <%- girModule.importNamespace -%>;