+++ packages/generator-typescript/templates/node-gtk/ambient.js
@@ -1,0 +1,5 @@
+<% if(moduleType === 'esm'){ %>
+    export {};
+<% } else { %>  
+    module.exports = {};
+<% } %>